pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "telemetry-studio"
version = "0.1.0"
description = "Web UI for telemetry video overlay configuration and preview"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
requires-python = ">=3.14"
authors = [
    { name = "Telemetry Studio Contributors" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "gopro-overlay>=0.129.0",
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "python-multipart>=0.0.22",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-playwright>=0.7.2",
    "httpx>=0.28.1",
    "ruff>=0.14.14",
]

[project.scripts]
telemetry-studio = "telemetry_studio.main:main"

[project.urls]
Homepage = "https://github.com/user/telemetry-studio"
Repository = "https://github.com/user/telemetry-studio"

[tool.hatch.build.targets.wheel]
packages = ["src/telemetry_studio"]

[tool.ruff]
target-version = "py314"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["telemetry_studio"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "e2e: marks tests as end-to-end UI tests (deselect with '-m \"not e2e\"')",
]
filterwarnings = [
    # Ignore deprecation warnings from third-party libraries
    "ignore:pkg_resources is deprecated:UserWarning",
    "ignore:websockets.legacy is deprecated:DeprecationWarning",
    "ignore:websockets.server.WebSocketServerProtocol is deprecated:DeprecationWarning",
    "ignore:'asyncio.iscoroutinefunction' is deprecated:DeprecationWarning",
    # Ignore unawaited coroutine warnings from AsyncMock in tests (expected when mocking async methods that may not be called)
    "ignore:coroutine 'AsyncMockMixin._execute_mock_call' was never awaited:RuntimeWarning",
]

[dependency-groups]
dev = [
    "nest-asyncio>=1.6.0",
]
